Created /hooks/useAuth to easily check if user is loggedin and handle logout.
